pyproject.toml
[project]
name = "SSParallaxX_bySamSeen"
description = "Create stunning parallax animations! Transform your ComfyUI generations into immersive 3D animations with realistic depth-based motion. Powered by advanced depth estimation, no external depth maps needed. Perfect for creating eye-catching content that leaps off the screen with realistic depth!"
version = "0.9.0"
license = "LICENSE"
dependencies = ["numpy", "Pillow", "torch", "transformers>=4.51.0", "opencv-python", "tqdm"]
keywords = ["parallax", "3D", "SS", "animation", "depth-map", "motion", "ComfyUI", "depth-estimation", "3D-animation", "depth-anything", "3D-video", "3D-motion", "gif", "video", "AI-depth", "immersive", "depth-based-motion"]

[project.urls]
Repository = "https://github.com/MrSamSeen/ComfyUI_SSStereoscope"
Documentation = "https://github.com/MrSamSeen/ComfyUI_SSStereoscope#readme"
Issues = "https://github.com/MrSamSeen/ComfyUI_SSStereoscope/issues"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "samseen"
DisplayName = "SSParallaxX by SamSeen"
Icon = ""
Tags = ["3D", "parallax", "depth-map", "animation", "video", "gif", "depth-anything"]
